Added SceneVisitController prefab
This is a prefab that has the script SceneVisitController script attached to it. All it does is when the scene is loaded, it updates Saves.json to reflect the progress the player has made, i.e., it will be noted that the last visited scene is the scene that just got loaded.