Tutorial Stages (Movement, Objects, and Buttons)
After abt 4 Unity crashes they are in a state of about 98% readiness. Last few issues involving the light switch for opening a door, rotating a door/wall, and having void object not show in mirror are all that is left I believe. Want to go over those in nxt meeting to get quickly fixed and then implement finished art assets